// frontend/src/api/generated/models/categoryResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Jeopareddy.Api
 * OpenAPI spec version: 1.0
 */
import type { ClueResponse } from './clueResponse';

export interface CategoryResponse {
  id?: string;
  /** @nullable */
  name?: string | null;
  displayOrder?: number;
  /** @nullable */
  clues?: ClueResponse[] | null;
}