// frontend/src/api/generated/models/clueResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Jeopareddy.Api
 * OpenAPI spec version: 1.0
 */

export interface ClueResponse {
  id?: string;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  answer?: string | null;
  /** @nullable */
  imageMimeType?: string | null;
  /** @nullable */
  imageBase64?: string | null;
  pointValue?: number;
  rowOrder?: number;
  isRevealed?: boolean;
  isAnswered?: boolean;
}