// frontend/src/api/generated/models/gameListItemResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Jeopareddy.Api
 * OpenAPI spec version: 1.0
 */

export interface GameListItemResponse {
  id?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: string | null;
  createdAtUtc?: string;
  updatedAtUtc?: string;
}