// frontend/src/api/generated/models/gameResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Jeopareddy.Api
 * OpenAPI spec version: 1.0
 */
import type { CategoryResponse } from './categoryResponse';
import type { TeamResponse } from './teamResponse';

export interface GameResponse {
  id?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: string | null;
  createdAtUtc?: string;
  updatedAtUtc?: string;
  /** @nullable */
  categories?: CategoryResponse[] | null;
  /** @nullable */
  teams?: TeamResponse[] | null;
}